+++ src/app/app.module.ts
@@ -1,4 +1,4 @@
-import { HttpClientModule } from '@angular/common/http';
+import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
@@ -12,10 +12,20 @@ import { GgcDatasetLegendModule } from 'generieke-geo-componenten-dataset-legend
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { DataService } from './services/data.service';
+import { AlertComponent } from './components/alert/alert.component';
+import { LoginComponent } from './login/login.component';
+import { JwtInterceptor } from './helpers/jwt.interceptor';
+import { ErrorInterceptor } from './helpers/error.interceptor';
+import { RegisterComponent } from './register/register.component';
+import { ViewerComponent } from './viewer/viewer.component';
 
 @NgModule({
   declarations: [
     AppComponent,
+    AlertComponent,
+    ViewerComponent,
+    LoginComponent,
+    RegisterComponent,
   ],
   imports: [
     BrowserModule,
@@ -28,7 +38,13 @@ import { DataService } from './services/data.service';
     GgcDatasetLegendModule,
     HttpClientModule,
   ],
-  providers: [DataService],
-  bootstrap: [AppComponent]
+  providers: [
+    DataService,
+    { provide: HTTP_INTERCEPTORS, useClass: JwtInterceptor, multi: true },
+    { provide: HTTP_INTERCEPTORS, useClass: ErrorInterceptor, multi: true },
+  ],
+  bootstrap: [
+    AppComponent,
+  ],
 })
 export class AppModule { }
