+++ src/app/components/alert/alert.component.html
@@ -1,0 +1,1 @@
+<div *ngIf="message" [ngClass]="message.cssClass">{{message.text}}</div>
