+++ src/app/helpers/error.interceptor.ts
@@ -1,0 +1,25 @@
+import { HttpEvent, HttpHandler, HttpInterceptor, HttpRequest } from '@angular/common/http';
+import { Injectable } from '@angular/core';
+import { Observable, throwError } from 'rxjs';
+import { catchError } from 'rxjs/operators';
+
+import { environment } from 'src/environments/environment';
+import { AuthenticationService } from '../services/authentication.service';
+
+@Injectable()
+export class ErrorInterceptor implements HttpInterceptor {
+    constructor(private authenticationService: AuthenticationService) {}
+
+    public intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
+        return next.handle(request).pipe(catchError((err) => {
+            if (err.status === 401 && !err.url.startsWith(environment.apiUrl)) {
+                // auto logout if 401 response returned from api
+                this.authenticationService.logout();
+                location.reload(true);
+            }
+
+            const error = err.error.message || err.statusText;
+            return throwError(error);
+        }));
+    }
+}
