+++ src/app/model/user.ts
@@ -1,0 +1,8 @@
+export class User {
+  id: number;
+  username: string;
+  password: string;
+  firstName: string;
+  lastName: string;
+  token: string;
+}
