+++ src/app/services/authentication.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { AuthenticationService } from './authentication.service';
+
+describe('AuthenticationService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: AuthenticationService = TestBed.get(AuthenticationService);
+    expect(service).toBeTruthy();
+  });
+});
